express definitions updated.
